Move query table related routines into a new module database/query-table.o
