+++ database/common.h
@@ -7,23 +7,28 @@
 *
 *  This file is copyright under the latest version of the EUPL.
 *  Please see LICENSE file for your rights under this license. */
-#ifndef DATABASE_H
-#define DATABASE_H
+#ifndef DATABASE_COMMON_H
+#define DATABASE_COMMON_H
 
+#include "sqlite3.h"
+
+bool check_database(int rc);
 void db_init(void);
 void *DB_thread(void *val);
-int get_number_of_queries_in_DB(void);
-void save_to_DB(void);
-void read_data_from_DB(void);
+int db_get_FTL_property(const unsigned int ID);
 bool db_set_FTL_property(const unsigned int ID, const int value);
 bool dbquery(const char *format, ...);
 bool dbopen(void);
 void dbclose(void);
 int db_query_int(const char*);
 void SQLite3LogCallback(void *pArg, int iErrCode, const char *zMsg);
+long int last_ID_in_DB(void);
+bool db_set_counter(const unsigned int ID, const int value);
+bool db_update_counters(const int total, const int blocked);
 
+extern sqlite3 *FTL_db;
 extern bool database;
 extern long int lastdbindex;
 extern bool DBdeleteoldqueries;
 
-#endif //DATABASE_H
+#endif //DATABASE_COMMON_H
